third_party/blink/web_tests/fast/forms/file/input-file-write-files-expected.txt

This tests the files attribute in file input forms

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS file1.files.length is 1
PASS file1.files = 'foo' threw exception TypeError: Failed to set the 'files' property on 'HTMLInputElement': Failed to convert value to 'FileList'..
PASS file1.files.length is 1
PASS file1.files.item(0).name is "foo.txt"
PASS file1.files.length is 1
PASS file1.files.item(0).name is "bar.txt"
PASS successfullyParsed is true

TEST COMPLETE

 
